+++ packages/ai/src/transports/transport-utils.ts
@@ -1,10 +1,9 @@
-import { createHash } from "node:crypto";
 import type { Model } from "@openclaw/llm-core";
 import { consumeResponseBytes } from "@openclaw/normalization-core";
 import { parseStrictPositiveInteger } from "@openclaw/normalization-core/number-coercion";
-import { normalizeOptionalString } from "@openclaw/normalization-core/string-coerce";
 import { truncateUtf16Safe } from "@openclaw/normalization-core/utf16-slice";
 import { getAiTransportHost } from "../host.js";
+export { redactIdentifier, sha256Hex } from "@openclaw/normalization-core/node-crypto";
 export { parseRetryAfterHeadersSeconds as parseRetryAfterSeconds } from "../internal/retry-after.js";
 
 export const MALFORMED_STREAMING_FRAGMENT_ERROR_MESSAGE =
@@ -19,23 +18,6 @@ export function parsePositiveInteger(value: unknown): number | undefined {
     return Math.floor(value);
   }
   return typeof value === "string" ? parseStrictPositiveInteger(value) : undefined;
-}
-
-export function sha256Hex(value: string | Uint8Array): string {
-  return createHash("sha256").update(value).digest("hex");
-}
-
-function sha256HexPrefix(value: string | Uint8Array, length: number): string {
-  return sha256Hex(value).slice(0, length);
-}
-
-export function redactIdentifier(value: string | undefined, opts?: { len?: number }): string {
-  const trimmed = normalizeOptionalString(value);
-  if (!trimmed) {
-    return "-";
-  }
-  const length = Number.isFinite(opts?.len) ? Math.max(1, Math.floor(opts?.len ?? 12)) : 12;
-  return `sha256:${sha256HexPrefix(trimmed, length)}`;
 }
 
 export function redactSensitiveText(text: string, _options?: unknown): string {
