+++ src/agents/auth-profiles/state-observation.ts
@@ -2,7 +2,7 @@
  * Structured logging for auth profile failure state changes.
  * Log payloads keep machine-readable fields while redacting console-facing ids.
  */
-import { redactIdentifier } from "../../logging/redact-identifier.js";
+import { redactIdentifier } from "@openclaw/normalization-core/node-crypto";
 import { createSubsystemLogger } from "../../logging/subsystem.js";
 import { sanitizeForConsole } from "../console-sanitize.js";
 import type { AuthProfileFailureReason, ProfileUsageStats } from "./types.js";
